small refactor to bank tab plugin; partly fix insertion
